backend/cloud_functions/onboard_patient/requirements.txt
# requirements.txt
functions-framework>=2.0.0
google-cloud-firestore>=2.0.0
google-cloud-secretmanager>=2.0.0
google-cloud-storage>=2.0.0
requests>=2.0.0